airflow/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
ffmpeg-python==0.2.0
future==1.0.0
minio==7.2.7
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pysrt==1.1.2
typing_extensions==4.12.2
urllib3==2.2.2
